Remove old MilestoneCard version completely.
